fixed functional and build errors from resolving merge conflicts
