+++ .github/workflows/deploy-frontend-to-prod.yaml
@@ -27,7 +27,7 @@ jobs:
               run: |
                 touch apps/frontend/.env.local
                 echo NEXT_PUBLIC_DYNAMIC_ENVIRONMENT_ID=08b64418-c698-4620-8ea2-d0e1a31ff10f >> apps/frontend/.env.local
-                echo API_ENDPOINT=https://api.aiden.space >> apps/frontend/.env.local
+                echo NEXT_PUBLIC_API_ENDPOINT=https://api.aiden.space >> apps/frontend/.env.local
                 echo AUTH_URL=https://aiden.space >> apps/frontend/.env.local
 
             - name: Build and tag frontend
