+++ apps/frontend/.env.local.example
@@ -5,7 +5,7 @@
 # Insensitive variables should have their actual values placed in here.
 
 # nodocker use http://localhost:8003, docker use http://host.docker.internal:8003
-API_ENDPOINT=
+NEXT_PUBLIC_API_ENDPOINT=
 # https://docs.dynamic.xyz/guides/frameworks/next-auth
 # NEXT_PUBLIC_DYNAMIC_ENVIRONMENT_ID loaded in GHA
 NEXT_PUBLIC_DYNAMIC_ENVIRONMENT_ID=08b64418-c698-4620-8ea2-d0e1a31ff10f
