+++ apps/frontend/lib/api/common.ts
@@ -1,3 +1,5 @@
+const API_ENDPOINT = process.env.API_ENDPOINT
+
 function fromApiEndpoint(url: string): URL {
   return new URL(url, process.env.NEXT_PUBLIC_API_ENDPOINT)
 }
@@ -45,7 +47,7 @@ async function createResource<T, P = undefined>(
     if (!response.ok)
       throw new Error(`Failed to create at ${baseUrl} with body ${body}`)
 
-    return response.json()
+    return response.json() as T
   } catch (error) {
     console.error(error)
   }
